Add photons_detected
This commit is intended to make it easier to do an 'apple to apple'
comparison between intensities of the original 'field' and of the
'deconvolved'.  Feedback is welcome.

In order to compare intensities of a deconvolved and original object I
changed the use of max_photons and added photons_detected.  The
simulation is now done as follows

1.  max_photons is used to control the number of photons emitted from
the field, so the final 'feild' = field*max_photons.  This way we can
compare the 'field' and deconvolved to see if intensities are
reconstructed correctly.  Previously max_photons was applied when
'poissrnd' was called, so photon count was reduced when the noise
function was applied.

2.  field is convolved with the OTF

3.  When noise is added we consider 'photons_detected', to compare
intensities of original field and deconvolved photons_detected should be
1.  Otherwise some photons are not 'detected' and the deconvolved will
less total intensity then the original.